cmma unit test: fix GL_INVALID_VALUE issue in glCopyTexSubImage2D
The issue was the w and h of the region to copy from was the size of the
screen instead of the size of the fbo, hence invalid value.

Adding some minor changes: remove fprintfs, increase loop count in the
rendering loop, change the filter of glBlitFramebuffer to NEAREST
and initialize gl_copy to avoid garbage values on subsequent re-runs.

Signed-off-by: Carlos Santa <carlos.santa@intel.com>